Implement starting configured system services.
Previously the entrypoint script provided utility functions for starting
runit services and waiting for ports to be open. This was useful when
the entrypoint script could be changed to start all required services.
This is no longer the case now and downstream containers will either
override the entrypoint script with a custom one or expect some kind of
default behavior. The default behavior implemented in this commit is to
start all configured system services as background jobs and quit as soon
as any of them exit or a signal (ex. interrupt) is received. This makes
it useful for interactive use as well.