Generate default system locale.
It turns out the runit-void package provides /etc/locale.conf which has
en_US.UTF-8 configured as the default locale for the system and many
other applications have this configured as a default in their init
scripts as well (ex. postgresql). So we need to install the
glibc-locales package, enable at least the default locale, and generate
it for these applications to work correctly.