Make wget less verbose.
The wget command to download the static XBPS tarball was unnecessarily
verbose. Use the --no-verbose flag to make it less verbose without
disabling all output (i.e. errors or warnings). Move the XBPS settings
higher in the build script to reflect the order in which the settings
are actually used.